+++ database/migrations/2024_09_06_151600_create_cart_items_table.php
@@ -4,17 +4,16 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
-    /**
-     * Run the migrations.
-     *
-     * @return void
-     */
-    public function up()
-    {
-        Schema::create('cart_items', function (Blueprint $table) {
-            $table->id();
+return new class extends Migration {
+	/**
+	 * Run the migrations.
+	 *
+	 * @return void
+	 */
+	public function up()
+	{
+		Schema::create('cart_items', function (Blueprint $table) {
+			$table->id();
 			$table->foreignId('cart_id')->constrained('carts')->cascadeOnUpdate()->cascadeOnDelete();
 			$table->string('title');
 			$table->foreignId('product_id')->constrained('products')->cascadeOnUpdate()->cascadeOnDelete();
@@ -27,17 +26,19 @@ return new class extends Migration
 			$table->decimal('width', 8, 2);
 			$table->decimal('height', 8, 2);
 			$table->decimal('length', 8, 2);
-            $table->timestamps();
-        });
-    }
+			$table->timestamps();
+		});
+	}
 
-    /**
-     * Reverse the migrations.
-     *
-     * @return void
-     */
-    public function down()
-    {
-        Schema::dropIfExists('cart_items');
-    }
+	/**
+	 * Reverse the migrations.
+	 *
+	 * @return void
+	 */
+	public function down()
+	{
+		if (app()->isLocal()) {
+			Schema::dropIfExists('cart_items');
+		}
+	}
 };
