+++ database/migrations/2024_10_05_151435_create_fabricators_table.php
@@ -4,30 +4,31 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
-    /**
-     * Run the migrations.
-     *
-     * @return void
-     */
-    public function up()
-    {
-        Schema::create('fabricators', function (Blueprint $table) {
-            $table->id();
-            $table->string('factory_title');
-            $table->string('country');
-            $table->timestamps();
-        });
-    }
+return new class extends Migration {
+	/**
+	 * Run the migrations.
+	 *
+	 * @return void
+	 */
+	public function up()
+	{
+		Schema::create('fabricators', function (Blueprint $table) {
+			$table->id();
+			$table->string('factory_title');
+			$table->string('country');
+			$table->timestamps();
+		});
+	}
 
-    /**
-     * Reverse the migrations.
-     *
-     * @return void
-     */
-    public function down()
-    {
-        Schema::dropIfExists('fabricators');
-    }
+	/**
+	 * Reverse the migrations.
+	 *
+	 * @return void
+	 */
+	public function down()
+	{
+		if (app()->isLocal()) {
+			Schema::dropIfExists('fabricators');
+		}
+	}
 };
