+++ database/migrations/2024_10_05_151449_create_promocodes_table.php
@@ -4,32 +4,33 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
-    /**
-     * Run the migrations.
-     *
-     * @return void
-     */
-    public function up()
-    {
-        Schema::create('promocodes', function (Blueprint $table) {
-            $table->id();
+return new class extends Migration {
+	/**
+	 * Run the migrations.
+	 *
+	 * @return void
+	 */
+	public function up()
+	{
+		Schema::create('promocodes', function (Blueprint $table) {
+			$table->id();
 			$table->string('itemable_type');
-            $table->string('name')->unique();
-            $table->integer('min_sum');
-            $table->integer('max_sum');
-            $table->timestamps();
-        });
-    }
+			$table->string('name')->unique();
+			$table->integer('min_sum');
+			$table->integer('max_sum');
+			$table->timestamps();
+		});
+	}
 
-    /**
-     * Reverse the migrations.
-     *
-     * @return void
-     */
-    public function down()
-    {
-        Schema::dropIfExists('promocodes');
-    }
+	/**
+	 * Reverse the migrations.
+	 *
+	 * @return void
+	 */
+	public function down()
+	{
+		if (app()->isLocal()) {
+			Schema::dropIfExists('promocodes');
+		}
+	}
 };
